backend/requirements.txt
a2wsgi==1.10.7
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
bentoml==1.3.9
blinker==1.8.2
cattrs==23.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
circus==0.18.0
click==8.1.7
click-option-group==0.5.6
cloudpickle==3.1.0
deepmerge==2.0
Deprecated==1.2.14
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.4.1
fs==2.4.16
fsspec==2024.9.0
gdown==5.2.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-ws==0.6.2
idna==3.10
importlib-metadata==6.11.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.1
numpy==2.1.2
nvidia-ml-py==11.525.150
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
pip-requirements-parser==32.0.1
prometheus_client==0.21.0
prompt-toolkit==3.0.36
propcache==0.2.0
psutil==6.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
questionary==2.0.1
requests==2.32.3
rich==13.9.2
schema==0.7.7
simple-di==0.1.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.0
sympy==1.13.3
tomli_w==1.1.0
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uv==0.4.22
uvicorn==0.32.0
watchfiles==0.24.0
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
yarl==1.15.4
zipp==3.20.2
